Use currrent working directory as default root path
Currently both the generator and validator use a hard-coded root path
when no path argument is given. This can be unexpected especially when
you have pds-skeleton installed globally.

Changing it to the cwd will mean the following are the same:

    pds-skeleton validate
    pds-skeleton validate .

    pds-skeleton generate
    pds-skeleton generate .

If getcwd() returns False then it defaults to the original hard-coded
path.